Save app config (just theme for the time being) to a file. Closes #23.
